separazione delle classi e aggiornamento service
