+++ Progetto_backend/src/main/java/com/puzzle_lab/entities/FasciaOraria.java
@@ -1,0 +1,5 @@
+package com.puzzle_lab.entities;
+
+public enum FasciaOraria {
+    MATTINA, POMERIGGIO
+}
