+++ Progetto_backend/src/main/java/com/puzzle_lab/entities/Laboratori.java
@@ -1,0 +1,14 @@
+package com.puzzle_lab.entities;
+
+public enum Laboratori {
+    STORIA_DI_CASCINA,
+    BRUNO_CACCIA,
+    INTRODUZIONE_AL_TEMA_MAFIA,
+    I_BENI_CONFISCATI,
+    LE_MAFIE_IN_PIEMONTE,
+    SOSTENIBILITA_AMBIENTALE,
+    LA_MAFIA_ATTRAVERSO_IL_CINEMA,
+    IL_GIOCO_NON_E_UN_AZZARDO,
+    CHI_NON_GIOCA_VINCE
+}
+
